requirements.txt
livekit==0.12.0
livekit-agents==0.5.0
google-generativeai==0.6.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
deepgram-sdk==3.2.7
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
